Fix run_tunnel fatal failure on FreeBSD
git-svn-id: https://svn.remlab.net/svn/miredo/trunk@1590 6046fb1f-dae2-0310-8472-869a43925389

+++ src/relay.c
@@ -387,7 +387,7 @@ run_tunnel (miredo_tunnel *tunnel, miredo_addrwatch *w)
 		return -1;
 
 	int fd = miredo_addrwatch_getfd (w);
-	if (fd >= FD_SETSIZE)
+	if (fd >= (int)FD_SETSIZE)
 		return -1;
 
 	/* Main loop */
